requirements.txt
fastapi==0.104.0
uvicorn==0.23.1
python-multipart==0.0.6
requests==2.31.0
faiss-cpu==1.9.0.post1
fastembed==0.7.4
groq
numpy
